+++ aws-sdk-core/spec/aws/instance_profile_credentials_spec.rb
@@ -88,6 +88,36 @@ module Aws
         expect(c.expiration.to_s).to eq(expiration2.to_s)
       end
 
+      it 'retries if get profile response is invalid JSON' do
+        stub_request(:get, "http://169.254.169.254#{path}").
+          to_return(:status => 500).
+          to_return(:status => 200, :body => "profile-name\n")
+        stub_request(:get, "http://169.254.169.254#{path}profile-name").
+          to_return(:status => 200, :body => ' ').
+          to_return(:status => 200, :body => '').
+          to_return(:status => 200, :body => '{').
+          to_return(:status => 200, :body => resp2)
+        c = InstanceProfileCredentials.new(backoff:0)
+        expect(c.credentials.access_key_id).to eq('akid-2')
+        expect(c.credentials.secret_access_key).to eq('secret-2')
+        expect(c.credentials.session_token).to eq('session-token-2')
+        expect(c.expiration.to_s).to eq(expiration2.to_s)
+      end
+
+      it 'retries invalid JSON exactly 3 times' do
+        stub_request(:get, "http://169.254.169.254#{path}").
+          to_return(:status => 500).
+          to_return(:status => 200, :body => "profile-name\n")
+        stub_request(:get, "http://169.254.169.254#{path}profile-name").
+          to_return(:status => 200, :body => '').
+          to_return(:status => 200, :body => ' ').
+          to_return(:status => 200, :body => '{').
+          to_return(:status => 200, :body => ' ')
+        expect {
+          InstanceProfileCredentials.new(backoff:0)
+        }.to raise_error(JSON::ParserError)
+      end
+
       describe 'auto refreshing' do
 
         # expire in 4 minutes